Append current state to power-sources-graph
